fix(F-SCR-001): Judgment.confidence 를 BigDecimal 로 — 결정 10.32 를 닫는다
ADR-008 이 금지한 double 이 해시 대상 한복판에 있었다. confidence 는 비율이고
ADR-008 문면이 "비율은 BigDecimal" 이다. #78 리뷰에서 1번(타입 변경)으로 정했고,
#86 이 그때까지 payload 변환으로 임시 다리를 놓아 뒀다 — 이 커밋으로 그 다리를 걷을 수 있다.

## 타입만 바꾸면 결정론이 되지 않는다

new BigDecimal(0.1) 은 이진 근사값을 그대로 옮겨 0.10000000000000000555… 가 된다. 그러면
같은 판정이 다른 바이트가 되고 해시가 갈리는데, 값을 눈으로 보면 0.1 이라 원인이 안 보인다.

확인한 것 — Jackson 은 목표 타입이 BigDecimal 이면 JSON 텍스트를 그대로 읽는다. 기본
설정으로도 0.91 이 정확히 온다. 그래서 AiServiceClient 의 경계 매퍼에는 손댈 것이 없다
(재생 경로는 Object 로 받아서 USE_BIG_DECIMAL_FOR_FLOATS 가 필요하다 — ImmutableStore).

ConfidencePrecisionTest 가 소스 전체를 훑어 new BigDecimal(<double 리터럴>) 이 없는 것을
단정한다. 목·픽스처가 double 리터럴을 쓰고 있어 거기서 새기 쉬운 종류라 테스트로 막는다.

## 게이트 R-05

Context.confidences 와 비교를 BigDecimal 로 옮겼다. 임계값도 Double.parseDouble 이 아니라
문자열에서 직접 만든다 — 0.7 이 이진 근사값이 되면 같은 "0.7" 을 비교하는데 판정이 표현
오차에 걸린다.

경계 동작은 안 바뀐다. 테스트로 고정했다.
  임계값과 정확히 같으면 발동하지 않는다 (0.7 → GREEN, 0.69 → YELLOW)
  자릿수만 다른 같은 신뢰도는 같게 판정된다 (0.70 ≡ 0.7)

## CanonicalJsonTest

"Judgment 는 confidence 가 double 이라 담을 수 없다" 테스트가 타입이 바뀌면 깨지도록
설계돼 있었고, 설계대로 깨졌다. 담긴다는 쪽으로 뒤집었다. 같은 모양의 대역 레코드
(JudgmentShaped)도 더 필요 없어 지우고 진짜 Judgment 를 쓴다.

계약(judgment.schema.json 의 type: number)과 web 은 안 바뀐다.

테스트 216건 통과.

+++ server/src/main/java/com/sphinxfin/sphinx/domain/Judgment.java
@@ -1,10 +1,12 @@
 package com.sphinxfin.sphinx.domain;
 
+import java.math.BigDecimal;
+
 /** contracts/judgment.schema.json 과 1:1. 근거(evidence) 없는 판정은 무효 (P4) */
 public record Judgment(
         String itemId,
         Grade grade,
-        double confidence,
+        BigDecimal confidence,
         Evidence evidence,
         String reason,
         String misconceptionType    // nullable, 예: M08-TYING
